[RISCV] Fix typo in CV_SH_rr_inc pattern
This typo caused a compiler crash in 'RISC-V Assembly Printer' because
CV_SH_ri_inc was selected, leading to `getImmOpValue` being called for
a register operand.

This bug did not affect the Assembler output and therefore does not trigger
any existing unit tests, but is visible by examining the final MIR function.